requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Supabase
supabase==2.4.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# AI / Speech
openai==1.30.1
pytesseract==0.3.10
Pillow==10.3.0

# PDF generation
reportlab==4.1.0

# WhatsApp
twilio==9.1.0

# Text-to-speech
gTTS==2.5.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
pydantic[email]
